pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "gitsync"
version = "0.1.0"
description = "Mirror GitLab activity into a GitHub repository for contribution tracking."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "Jakub Krasuski" }
]
dependencies = [
  "aiosqlite>=0.21.0,<1.0.0",
  "fastapi>=0.115.0,<1.0.0",
  "httpx>=0.28.0,<1.0.0",
  "jinja2>=3.1.0,<4.0.0",
  "pydantic-settings>=2.8.0,<3.0.0",
  "python-multipart>=0.0.20,<1.0.0",
  "typer>=0.16.0,<1.0.0",
  "uvicorn[standard]>=0.34.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "black>=25.1.0,<26.0.0",
  "mypy>=1.15.0,<2.0.0",
  "pytest>=8.3.0,<9.0.0",
  "pytest-asyncio>=0.26.0,<1.0.0",
  "pytest-cov>=6.1.0,<7.0.0",
  "ruff>=0.11.0,<1.0.0",
]

[project.scripts]
gitsync = "gitsync.cli:app"

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra --strict-markers --cov=src/gitsync --cov-report=term-missing"
asyncio_mode = "auto"
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "C4", "W"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["B011"]

[tool.mypy]
python_version = "3.11"
disallow_untyped_defs = true
disallow_incomplete_defs = true
warn_return_any = true
warn_unused_ignores = true
show_error_codes = true
strict_optional = true
packages = ["gitsync"]

[tool.hatch.build.targets.wheel]
packages = ["src/gitsync"]